Menambahkan fitur untuk menyimpan kategori baru dalam BarangController dan memperbarui validasi input. Mengubah tipe data harga dan modal menjadi bigInteger di migrasi. Memperbarui rute API untuk penanganan stok barang dengan nama yang lebih deskriptif.

+++ database/migrations/2025_08_12_030241_create_barangs_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('nama');
             $table->string('kategori');
             $table->integer('stok');
-            $table->double('harga', 15,2);  
-            $table->double('modal', 15, 2);
+            $table->bigInteger('harga');  
+            $table->bigInteger('modal');
             $table->string('barcode')->nullable();
             $table->string('gambar_path')->nullable();
             $table->timestamps();
